upgrade script enahncements and now temporary in test mode

+++ src/opnsense/scripts/pkg_upgrade.sh
@@ -1,5 +1,17 @@
 #!/bin/sh
-pkg upgrade -y > /tmp/pkg_upgrade.progress &
-echo '***DONE***' >> /tmp/pkg_upgrade.progress
-rm /tmp/pkg_upgrade.progress
+
+pkg_running=`ps -x | grep "pkg " | grep -v "grep"`
+if [ "$pkg_running" == "" ]; then
+	if [ -f /tmp/pkg_upgrade.progress ]
+		# Remove leftovers from previous upgrade first
+		rm /tmp/pkg_upgrade.progress
+	fi
+      # start pkg upgrade
+	pkg upgrade -n > /tmp/pkg_upgrade.progress & # Need to st this to -y for production, now set to -n for testing purpose
+	echo '***DONE***' >> /tmp/pkg_upgrade.progress
+else
+	echo 'Upgrade already in progress'
+	echo '***DONE***'
+fi
+
